+++ ppa/archive/management/commands/gale_import.py
@@ -28,25 +28,20 @@ These are the supported collection abbreviations:
 
 """
 import csv
-import json
 import logging
-import os.path
-import time
 from collections import Counter
 
-import pymarc
 from django.conf import settings
 from django.contrib.admin.models import ADDITION, LogEntry
 from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ImproperlyConfigured
 from django.core.management.base import BaseCommand, CommandError
-from django.db.utils import IntegrityError
 from django.template.defaultfilters import pluralize, truncatechars
-from pairtree import PairtreeStorageFactory
 from parasolr.django.signals import IndexableSignalHandler
 
 from ppa.archive.gale import GaleAPI, GaleAPIError, MARCRecordNotFound, get_marc_record
+from ppa.archive.import_util import GaleImporter
 from ppa.archive.models import Collection, DigitizedWork, Page
 
 logger = logging.getLogger(__name__)
@@ -74,15 +69,12 @@ class Command(BaseCommand):
         "WL": "Word Lists",
     }
 
-    # path to gale id lookup file
-    gale_id_path = os.path.join(
-        settings.BASE_DIR, "ppa", "archive", "fixtures", "gale_id_lookup.json"
-    )
-    # NOTE: as of June 2021, Gale API item details does not include the ESTC id needed
-    # for accessing the correct MARC record, nor does it include volume information.
-    # For now, we use a lookup file generated from a spreadsheet provided by Gale
-    # for the records we plan to import. We will revisit this when the information
-    # is made available through their API.
+    # item type lookup for supported types — adapted from hathi excerpt script
+    item_type = {
+        "Excerpt": DigitizedWork.EXCERPT,
+        "Article": DigitizedWork.ARTICLE,
+        "Full work": DigitizedWork.FULL,
+    }
 
     def add_arguments(self, parser):
         parser.add_argument(
@@ -127,11 +119,6 @@ class Command(BaseCommand):
             raise CommandError(str(err))
 
         self.stats = Counter()
-        self.script_user = User.objects.get(username=settings.SCRIPT_USERNAME)
-        self.digwork_contentype = ContentType.objects.get_for_model(DigitizedWork)
-        # load json gale id lookup file so we can get ESTC ids from document id
-        with open(self.gale_id_path) as idfile:
-            self.id_lookup = json.load(idfile)
 
         # if ids are specified on the command line, create a list
         # of dictionaries so import will look similar to csv
@@ -146,6 +133,10 @@ class Command(BaseCommand):
         # total is needed for progessbar (if we add it)
         self.stats["total"] = len(to_import)
 
+        # initialize importer and run pre-steps needed before importi
+        self.importer = GaleImporter()
+        self.importer.add_item_prep()
+
         for item in to_import:
             if self.verbosity >= self.v_normal:
                 # include title in output if present, but truncate since many are long
@@ -156,7 +147,7 @@ class Command(BaseCommand):
             # to handle notes and collection membership from CSV
             item_info = item.copy()
             del item_info["ID"]  # don't send ID twice
-            self.import_digitizedwork(item["ID"], **item_info)
+            self.import_record(item["ID"], **item_info)
 
         summary = (
             "\nProcessed {:,d} item{} for import."
@@ -199,81 +190,57 @@ class Command(BaseCommand):
             raise CommandError("ID column is required in CSV file")
         return data
 
-    def import_digitizedwork(self, gale_id, **kwargs):
+    def import_record(self, gale_id, **kwargs):
         """Import a single work into the database.
-        Retrieves bibliographic data from Gale API."""
+        Retrieves record data from Gale API."""
 
-        # if an item with this source id exists, skip
+        # check if an item with this source id + page range exists
         # (check local db first because API call is slow for large items)
-        # NOTE: revisit if we decide to support update logic
-        if DigitizedWork.objects.filter(source_id=gale_id).exists():
+
+        # use an unsaved digitized work to parse the page range (if any)
+        # for queryset filter to check for duplicates
+        dw_pages = DigitizedWork(
+            pages_digital=kwargs.get("Digital Page Range", "").replace(";", ",")
+        )
+
+        if DigitizedWork.objects.filter(
+            source_id=gale_id, pages_digital=dw_pages.pages_digital
+        ).exists():
             self.stderr.write("%s is already in the database; skipping" % gale_id)
             self.stats["skipped"] += 1
             return
 
-        try:
-            item_record = self.gale_api.get_item(gale_id)
-        except GaleAPIError as err:
-            self.stderr.write("Error getting item information for %s" % gale_id)
-            self.stats["error"] += 1
-            return
-
-        # document metadata is under "doc"
-        doc_metadata = item_record["doc"]
-        # NOTE: url provided in "isShownAt" includes user and source parameters;
-        # these are important to preserve because it allows Gale to
-        # monitor traffic source and the linked page is an "auth free" link
-
-        # create new stub record and populate it from api response
-        digwork = DigitizedWork(
-            source_id=gale_id,  # or doc_metadata['id']; format CW###
-            source=DigitizedWork.GALE,
-            record_id=self.id_lookup[gale_id]["estc_id"],
-            source_url=doc_metadata["isShownAt"],
-            enumcron=self.id_lookup[gale_id].get("volume", ""),
-            title=doc_metadata["title"],
-            page_count=len(item_record["pageResponse"]["pages"]),
-            # import any notes from csv as private notes
-            notes=kwargs.get("NOTES", ""),
-        )
-        # populate titles, author, publication info from marc record
-        try:
-            digwork.metadata_from_marc(get_marc_record(digwork.record_id))
-        except MARCRecordNotFound:
-            self.stats["no_marc"] += 1
-            self.stderr.write(
-                self.style.WARNING(
-                    "MARC record not found for %s/%s" % (gale_id, digwork.record_id)
-                )
-            )
-        digwork.save()
-
-        # set collection membership based on spreadsheet columns
+        # determine collection membership based on spreadsheet columns
         digwork_collections = [
             collection
             for code, collection in self.collections.items()
             if kwargs.get(code)
         ]
-        if digwork_collections:
-            digwork.collections.set(digwork_collections)
 
-        # create log entry to document import
-        LogEntry.objects.log_action(
-            user_id=self.script_user.pk,
-            content_type_id=self.digwork_contentype.pk,
-            object_id=digwork.pk,
-            object_repr=str(digwork),
-            change_message="Created from Gale API",
-            action_flag=ADDITION,
+        # translate item type in spreadsheet to digitized work item type code
+        # strip whitespace in case any was added in the spreadsheet
+        kwargs["item_type"] = self.item_type.get(kwargs.get("Item Type", "").strip())
+
+        digwork = self.importer.import_digitizedwork(
+            gale_id, collections=digwork_collections, **kwargs
         )
 
-        # index the work once (signals index twice becuase of m2m change)
-        DigitizedWork.index_items([digwork])
+        # if import failed, check status
+        if not digwork:
+            if isinstance(self.importer.results[gale_id], GaleAPIError):
+                self.stderr.write("Error getting item information for %s" % gale_id)
+                self.stats["error"] += 1
+                return
 
-        # item record used for import includes page metadata;
-        # for efficiency, index pages at import time with the same api response
-        DigitizedWork.index_items(Page.gale_page_index_data(digwork, item_record))
+        # check for marc record not found error
+        if isinstance(self.importer.results[gale_id], MARCRecordNotFound):
+            self.stats["no_marc"] += 1
+            self.stderr.write(
+                self.style.WARNING("MARC record not found for %s" % (gale_id))
+            )
 
-        self.stats["imported"] += 1
-        self.stats["pages"] += digwork.page_count
-        return digwork
+        # if record was created successfully, update stats
+        if digwork:
+            self.stats["imported"] += 1
+            self.stats["pages"] += digwork.page_count
+            return digwork